Remove custom Accept header for hide comment preview API
This is not part of the main API. This is to help test https://github.com/runatlantis/atlantis/issues/1009 For some reason API calls come back with '406 Not Acceptable Body' on Github Enterprise, despite the GE stating the custom Accept headers are required.

Maybe it's no longer required like on github.com and GH simply forgot to update the docs?